# Filter SMT lemmas for predicate checks and get-model