Related: #4034 

Booster SMT bindings change as follows:

* SMT lemmas are held in a lookup table separate from the sort and function declarations
* The lookup is by `SymbolName`, and returns lemmas that (transitively) _may_ relate to the given symbol
* When checking predicates or computing a model, only those lemmas are added which relate to a symbol used in one of the predicates

This is hopefully reducing load on the SMT solver to enable solving problems that came up in engagement proofs and led to timeouts.