@jberthold 2024-09-11
`evm-semantics` tests with a regression in `sum-to-n-spec.k`

| Test                     | 4043-filter-smt-lemmas-for-predicate-checks time | master-960328a51 time | (4043-filter-smt-lemmas-for-predicate-checks/master-960328a51) time 
|--------------------------|--------------------------------------------------|-----------------------|---------------------------------------------------------------------
| examples/sum-to-n-spec.k | 96.15                                            | 70.09                 | 1.3718076758453417                                                  
| TOTAL                    | 96.15                                            | 70.09                 | 1.3718076758453417