@geo2a 2024-09-26
KEVM performance at [4e78fe8](https://github.com/runtimeverification/haskell-backend/pull/4048/commits/4e78fe8422e75ec56aae586a96abd817ec634e3f)
| Test                                           | 4043-filter-smt-lemmas-for-predicate-checks time | master-79caf8c55 time | (4043-filter-smt-lemmas-for-predicate-checks/master-79caf8c55) time
|------------------------------------------------|--------------------------------------------------|-----------------------|---------------------------------------------------------------------
| mcd/dsvalue-read-pass-spec.k                   | 60.21                                            | 62.7                  | 0.9602870813397129
| benchmarks/storagevar01-spec.k                 | 49.31                                            | 47.19                 | 1.0449247721974995
| mcd-structured/pot-join-pass-rough-spec.k      | 201.05                                           | 182.56                | 1.1012817703768625
| examples/erc20-spec.md                         | 171.51                                           | 154.13                | 1.112761954194511
| benchmarks/ecrecoverloop02-sig1-invalid-spec.k | 118.25                                           | 99.96                 | 1.1829731892757103
| erc20/ds/transfer-success-1-spec.k             | 75.74                                            | 58.13                 | 1.3029416824359192
| examples/sum-to-n-spec.k                       | 134.98                                           | 100.83                | 1.3386888822771
| TOTAL                                          | 811.0500000000001                                | 705.5                 | 1.1496102055279944